refactor(SessionPrice): price 인스턴스 변수 타입 변경(int -> long)

+++ src/main/java/nextstep/courses/domain/SessionChargeType.java
@@ -3,9 +3,9 @@ package nextstep.courses.domain;
 public enum SessionChargeType {
     FREE, CHARGED;
 
-    private static final int FREE_OF_CHARGE = 0;
+    private static final long FREE_OF_CHARGE = 0L;
 
-    public static SessionChargeType of(int price) {
+    public static SessionChargeType of(long price) {
         if(price == FREE_OF_CHARGE) {
             return FREE;
         }
